+++ changes/logic.cpp
@@ -143,7 +143,7 @@ void changes_clear_notifications_user (string jobid, string username)
     cleared_count_in_one_go = database_modifications.clearNotificationsUser (username);
     total_cleared += cleared_count_in_one_go;
     if (!identifiers.empty ()) {
-      database_jobs.set_percentage (convert_to_int (jobid), 100 * total_cleared / identifiers.size());
+      database_jobs.set_percentage (convert_to_int (jobid), 100 * total_cleared / static_cast<int> (identifiers.size()));
     }
   } while (cleared_count_in_one_go);
