+++ notes/notes.cpp
@@ -94,12 +94,12 @@ string notes_notes (void * webserver_request)
         numeric_passages.push_back (filter_passage_to_integer (passage));
       }
       if (!numeric_passages.empty ()) {
-        double average = accumulate (numeric_passages.begin (), numeric_passages.end (), 0) / numeric_passages.size ();
-        passage_sort_key = round(average);
+        double average = static_cast<size_t>(accumulate (numeric_passages.begin (), numeric_passages.end (), 0)) / numeric_passages.size ();
+        passage_sort_key = static_cast<int> (round(average));
       }
       passage_sort_keys.push_back (passage_sort_key);
     }
-    quick_sort (passage_sort_keys, identifiers, 0, (int)identifiers.size ());
+    quick_sort (passage_sort_keys, identifiers, 0, identifiers.size ());
   }
